refactor(profile): split useProfile into modular hooks
Break useProfile into focused profile modules (CRUD, autosave,
history, and pure helpers) while preserving the existing public hook
contract and IPC behavior. This reduces file sizes below the 500-line
policy threshold for task 7 under issue #290.
