+++ src/crosshook-native/src/hooks/profile/constants.ts
@@ -1,0 +1,2 @@
+export const automaticLauncherSuffix = ' - Trainer';
+export const launchOptimizationsAutosaveDelayMs = 350;
