+++ src/crosshook-native/src/hooks/profile/launchOptimizationIds.ts
@@ -1,0 +1,43 @@
+import type { LaunchOptimizationId } from '../../types/launch-optimizations';
+import type { OptimizationEntry } from '../../utils/optimization-catalog';
+
+export function normalizeLaunchOptimizationIds(
+  ids: readonly string[] | undefined,
+  optionsById: Record<string, OptimizationEntry>
+): LaunchOptimizationId[] {
+  if (ids === undefined) {
+    return [];
+  }
+
+  const catalogLoaded = Object.keys(optionsById).length > 0;
+  const normalized: LaunchOptimizationId[] = [];
+  const seenIds = new Set<LaunchOptimizationId>();
+
+  for (const optionId of ids) {
+    // When catalog is not yet loaded, pass IDs through without filtering (lenient mode).
+    if (catalogLoaded && !(optionId in optionsById)) {
+      continue;
+    }
+
+    const typedOptionId = optionId as LaunchOptimizationId;
+    if (seenIds.has(typedOptionId)) {
+      continue;
+    }
+
+    seenIds.add(typedOptionId);
+    normalized.push(typedOptionId);
+  }
+
+  return normalized;
+}
+
+export function areLaunchOptimizationIdsEqual(
+  left: readonly LaunchOptimizationId[],
+  right: readonly LaunchOptimizationId[]
+): boolean {
+  if (left.length !== right.length) {
+    return false;
+  }
+
+  return left.every((optionId, index) => optionId === right[index]);
+}
