+++ src/crosshook-native/src/hooks/profile/mergeRecentPaths.ts
@@ -1,0 +1,8 @@
+export function mergeRecentPaths(currentPaths: string[], nextPath: string): string[] {
+  const trimmed = nextPath.trim();
+  if (!trimmed) {
+    return currentPaths;
+  }
+
+  return [trimmed, ...currentPaths.filter((path) => path !== trimmed)].slice(0, 10);
+}
