+++ src/crosshook-native/src/hooks/useProfile.ts
@@ -1,43 +1,31 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useEffect, useMemo, useRef } from 'react';
 import { subscribeEvent } from '@/lib/events';
-import { callCommand } from '@/lib/ipc';
 import type {
-  AppSettingsData,
   BundledOptimizationPreset,
   ConfigDiffResult,
   ConfigRevisionSummary,
   ConfigRollbackResult,
   GameProfile,
   LaunchAutoSaveStatus,
-  LauncherInfo,
-  RecentFilesData,
-  SerializedDuplicateProfileResult,
-  SerializedGameProfile,
 } from '../types';
+import type { LaunchOptimizationId } from '../types/launch-optimizations';
+import { buildConflictMatrix, buildOptionsById, type OptimizationCatalogPayload } from '../utils/optimization-catalog';
+import type { LaunchOptimizationsStatus as LaunchOptimizationsStatusType } from './profile/launchOptimizationStatus';
 import {
-  getConflictingLaunchOptimizationIds,
-  type LaunchOptimizationId,
-  type LaunchOptimizations,
-} from '../types/launch-optimizations';
-import { DEFAULT_GAMESCOPE_CONFIG, DEFAULT_MANGOHUD_CONFIG, normalizeSerializedGameProfile } from '../types/profile';
-import { toSettingsSaveRequest } from '../types/settings';
-import { type ResolvedLaunchMethod, resolveLaunchMethod } from '../utils/launch';
-import {
-  buildConflictMatrix,
-  buildOptionsById,
-  type OptimizationCatalogPayload,
-  type OptimizationEntry,
-} from '../utils/optimization-catalog';
+  type PendingDelete as PendingDeleteType,
+  type PersistProfileDraftResult as PersistProfileDraftResultType,
+  type PersistProfileDraft as PersistProfileDraftType,
+  type RenameProfileResult as RenameProfileResultType,
+  useProfileCrud,
+} from './profile/useProfileCrud';
+import { useProfileHistory } from './profile/useProfileHistory';
+import { useProfileLaunchAutosave } from './profile/useProfileLaunchAutosave';
 import { useLaunchOptimizationCatalog } from './useLaunchOptimizationCatalog';
 
-export interface PendingDelete {
-  name: string;
-  launcherInfo: LauncherInfo | null;
-}
-
-export type PersistProfileDraftResult = { ok: true } | { ok: false; error: string };
-
-export type PersistProfileDraft = (name: string, profile: GameProfile) => Promise<PersistProfileDraftResult>;
+export type PendingDelete = PendingDeleteType;
+export type PersistProfileDraftResult = PersistProfileDraftResultType;
+export type PersistProfileDraft = PersistProfileDraftType;
+export type RenameProfileResult = RenameProfileResultType;
 
 export interface UseProfileResult {
   profiles: string[];
@@ -130,1225 +118,55 @@ export interface UseProfileOptions {
   onAfterRollback?: (profileName: string) => void;
 }
 
-export interface RenameProfileResult {
-  ok: boolean;
-  hadLauncher: boolean;
-}
-
-const automaticLauncherSuffix = ' - Trainer';
-const launchOptimizationsAutosaveDelayMs = 350;
-
-/** Tauri invoke failures are sometimes plain objects, not Error instances. */
-function formatInvokeError(err: unknown): string {
-  if (err instanceof Error) {
-    return err.message;
-  }
-  if (typeof err === 'string') {
-    return err;
-  }
-  if (err && typeof err === 'object') {
-    const message = (err as { message?: unknown }).message;
-    if (typeof message === 'string' && message.length > 0) {
-      return message;
-    }
-  }
-  try {
-    return JSON.stringify(err);
-  } catch {
-    return String(err);
-  }
-}
-
 /** @deprecated Use `LaunchAutoSaveStatus` from `../types` instead. */
 export type LaunchOptimizationsStatusTone = LaunchAutoSaveStatus['tone'];
 
 /** @deprecated Use `LaunchAutoSaveStatus` from `../types` instead. */
-export type LaunchOptimizationsStatus = LaunchAutoSaveStatus;
-
-function stripAutomaticLauncherSuffix(value: string): string {
-  const trimmed = value.trim();
-  return trimmed.endsWith(automaticLauncherSuffix)
-    ? trimmed.slice(0, -automaticLauncherSuffix.length).trimEnd()
-    : trimmed;
-}
-
-function deriveDisplayNameFromPath(path: string): string {
-  const normalized = path.trim();
-  if (!normalized) {
-    return '';
-  }
-
-  const segment = normalized.split(/[\\/]/).pop() ?? '';
-  return segment.replace(/\.[^.]+$/, '').trim();
-}
-
-function deriveGameName(profile: GameProfile): string {
-  return profile.game.name.trim() || deriveDisplayNameFromPath(profile.game.executable_path);
-}
-
-function deriveLauncherDisplayName(profile: GameProfile): string {
-  return (
-    stripAutomaticLauncherSuffix(profile.steam.launcher.display_name) ||
-    deriveGameName(profile) ||
-    stripAutomaticLauncherSuffix(deriveDisplayNameFromPath(profile.trainer.path))
-  );
-}
-
-function normalizeLaunchOptimizationIds(
-  ids: readonly string[] | undefined,
-  optionsById: Record<string, OptimizationEntry>
-): LaunchOptimizationId[] {
-  if (ids === undefined) {
-    return [];
-  }
-
-  const catalogLoaded = Object.keys(optionsById).length > 0;
-  const normalized: LaunchOptimizationId[] = [];
-  const seenIds = new Set<LaunchOptimizationId>();
-
-  for (const optionId of ids) {
-    // When catalog is not yet loaded, pass IDs through without filtering (lenient mode).
-    if (catalogLoaded && !(optionId in optionsById)) {
-      continue;
-    }
-
-    const typedOptionId = optionId as LaunchOptimizationId;
-    if (seenIds.has(typedOptionId)) {
-      continue;
-    }
-
-    seenIds.add(typedOptionId);
-    normalized.push(typedOptionId);
-  }
-
-  return normalized;
-}
-
-type ApplyLaunchOptimizationToggleResult = { ok: true; profile: GameProfile } | { ok: false; conflictLabels: string[] };
-
-function applyLaunchOptimizationToggle(
-  current: GameProfile,
-  optionId: LaunchOptimizationId,
-  nextEnabled: boolean,
-  optionsById: Record<string, OptimizationEntry>,
-  conflictMatrix: Readonly<Record<string, readonly string[]>>
-): ApplyLaunchOptimizationToggleResult {
-  const currentIds = current.launch.optimizations.enabled_option_ids;
-  const conflictingIds = nextEnabled ? getConflictingLaunchOptimizationIds(optionId, currentIds, conflictMatrix) : [];
-
-  if (conflictingIds.length > 0) {
-    const conflictLabels = conflictingIds.map((conflictingId) => optionsById[conflictingId]?.label ?? conflictingId);
-    return { ok: false, conflictLabels };
-  }
-
-  const nextIds = nextEnabled
-    ? normalizeLaunchOptimizationIds([...currentIds, optionId], optionsById)
-    : currentIds.filter((currentOptionId) => currentOptionId !== optionId);
-
-  const activeKey = (current.launch.active_preset ?? '').trim();
-  const presets = { ...(current.launch.presets ?? {}) };
-  if (activeKey && presets[activeKey]) {
-    presets[activeKey] = { enabled_option_ids: nextIds };
-  }
-
-  return {
-    ok: true,
-    profile: {
-      ...current,
-      launch: {
-        ...current.launch,
-        presets,
-        optimizations: {
-          enabled_option_ids: nextIds,
-        },
-      },
-    },
-  };
-}
-
-function areLaunchOptimizationIdsEqual(
-  left: readonly LaunchOptimizationId[],
-  right: readonly LaunchOptimizationId[]
-): boolean {
-  if (left.length !== right.length) {
-    return false;
-  }
-
-  return left.every((optionId, index) => optionId === right[index]);
-}
-
-function buildLaunchOptimizationsStatus(
-  method: ResolvedLaunchMethod,
-  hasExistingSavedProfile: boolean
-): LaunchOptimizationsStatus {
-  if (method !== 'proton_run' && method !== 'steam_applaunch') {
-    return {
-      tone: 'warning',
-      label: 'Unavailable for current method',
-      detail: 'Launch optimizations are only editable when the profile method is proton_run or steam_applaunch.',
-    };
-  }
-
-  if (!hasExistingSavedProfile) {
-    return {
-      tone: 'warning',
-      label: 'Save profile first to enable autosave',
-      detail: 'Optimization changes stay local until the profile has been saved once.',
-    };
-  }
-
-  return {
-    tone: 'idle',
-    label: 'Ready to autosave',
-    detail:
-      method === 'steam_applaunch'
-        ? 'Only launch.optimizations will be written automatically; paste the generated line into Steam yourself.'
-        : 'Only launch.optimizations will be written automatically for this saved profile.',
-  };
-}
-
-function normalizeLaunchPresetsSection(
-  profile: GameProfile,
-  optionsById: Record<string, OptimizationEntry>
-): {
-  presets: Record<string, LaunchOptimizations>;
-  active_preset: string;
-} {
-  const raw = profile.launch.presets;
-  const presets: Record<string, LaunchOptimizations> = {};
-  if (raw && typeof raw === 'object') {
-    for (const [key, value] of Object.entries(raw)) {
-      const name = key.trim();
-      if (!name) {
-        continue;
-      }
-      presets[name] = {
-        enabled_option_ids: normalizeLaunchOptimizationIds(value?.enabled_option_ids, optionsById),
-      };
-    }
-  }
-  const active_preset = (profile.launch.active_preset ?? '').trim();
-  return { presets, active_preset };
-}
-
-function normalizeProfileForEdit(
-  profile: SerializedGameProfile | GameProfile,
-  optionsById: Record<string, OptimizationEntry>
-): GameProfile {
-  const normalizedProfile = normalizeSerializedGameProfile(profile);
-  const method = resolveLaunchMethod(normalizedProfile);
-  const runtime = normalizedProfile.runtime;
-  const { presets, active_preset } = normalizeLaunchPresetsSection(normalizedProfile, optionsById);
-  let enabledOptionIds = normalizeLaunchOptimizationIds(
-    normalizedProfile.launch.optimizations?.enabled_option_ids,
-    optionsById
-  );
-  if (active_preset && presets[active_preset]) {
-    enabledOptionIds = presets[active_preset].enabled_option_ids;
-  }
-
-  return {
-    ...normalizedProfile,
-    trainer: {
-      ...normalizedProfile.trainer,
-      type: normalizedProfile.trainer.type.trim(),
-      loading_mode: normalizedProfile.trainer.loading_mode ?? 'source_directory',
-    },
-    steam: {
-      ...normalizedProfile.steam,
-      enabled: method === 'steam_applaunch',
-      launcher: {
-        ...normalizedProfile.steam.launcher,
-        display_name: stripAutomaticLauncherSuffix(normalizedProfile.steam.launcher.display_name),
-      },
-    },
-    runtime: {
-      ...runtime,
-      prefix_path: runtime.prefix_path.trim(),
-      proton_path: runtime.proton_path.trim(),
-      working_directory: runtime.working_directory.trim(),
-    },
-    launch: {
-      ...normalizedProfile.launch,
-      method,
-      presets,
-      active_preset,
-      optimizations: {
-        enabled_option_ids: enabledOptionIds,
-      },
-      custom_env_vars: { ...(normalizedProfile.launch.custom_env_vars ?? {}) },
-    },
-  };
-}
-
-function normalizeProfileForSave(profile: GameProfile, optionsById: Record<string, OptimizationEntry>): GameProfile {
-  const normalized = normalizeProfileForEdit(profile, optionsById);
-
-  return {
-    ...normalized,
-    game: {
-      ...normalized.game,
-      name: deriveGameName(normalized),
-    },
-    trainer: {
-      ...normalized.trainer,
-    },
-    steam: {
-      ...normalized.steam,
-      launcher: {
-        ...normalized.steam.launcher,
-        display_name: deriveLauncherDisplayName(normalized),
-      },
-    },
-  };
-}
-
-function validateProfileForSave(profile: GameProfile): string | null {
-  if (!profile.game.executable_path.trim()) {
-    return 'Game executable path is required before saving a profile.';
-  }
-
-  return null;
-}
-
-function mergeRecentPaths(currentPaths: string[], nextPath: string): string[] {
-  const trimmed = nextPath.trim();
-  if (!trimmed) {
-    return currentPaths;
-  }
-
-  return [trimmed, ...currentPaths.filter((path) => path !== trimmed)].slice(0, 10);
-}
-
-function createEmptyProfile(): GameProfile {
-  return {
-    game: {
-      name: '',
-      executable_path: '',
-    },
-    trainer: {
-      path: '',
-      type: '',
-      loading_mode: 'source_directory',
-    },
-    injection: {
-      dll_paths: [],
-      inject_on_launch: [false, false],
-    },
-    steam: {
-      enabled: false,
-      app_id: '',
-      compatdata_path: '',
-      proton_path: '',
-      launcher: {
-        icon_path: '',
-        display_name: '',
-      },
-    },
-    runtime: {
-      prefix_path: '',
-      proton_path: '',
-      working_directory: '',
-    },
-    launch: {
-      method: 'proton_run',
-      optimizations: {
-        enabled_option_ids: [],
-      },
-      presets: {},
-      active_preset: '',
-      custom_env_vars: {},
-    },
-  };
-}
+export type LaunchOptimizationsStatus = LaunchOptimizationsStatusType;
 
 export function useProfile(options: UseProfileOptions = {}): UseProfileResult {
-  const [profiles, setProfiles] = useState<string[]>([]);
-  const [favoriteProfiles, setFavoriteProfiles] = useState<string[]>([]);
-  const [selectedProfile, setSelectedProfile] = useState('');
-  const [profileName, setProfileName] = useState('');
-  const [profile, setProfile] = useState<GameProfile>(createEmptyProfile);
-  const [dirty, setDirty] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [saving, setSaving] = useState(false);
-  const [deleting, setDeleting] = useState(false);
-  const [duplicating, setDuplicating] = useState(false);
-  const [renaming, setRenaming] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [historyLoading, setHistoryLoading] = useState(false);
-  const [historyError, setHistoryError] = useState<string | null>(null);
-  const [pendingDelete, setPendingDelete] = useState<PendingDelete | null>(null);
-  const [launchOptimizationsStatus, setLaunchOptimizationsStatus] = useState<LaunchOptimizationsStatus>(
-    buildLaunchOptimizationsStatus('proton_run', false)
-  );
-  const [bundledOptimizationPresets, setBundledOptimizationPresets] = useState<BundledOptimizationPreset[]>([]);
-  const [optimizationPresetActionBusy, setOptimizationPresetActionBusy] = useState(false);
-  const launchOptimizationsAutosaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const lastSavedLaunchOptimizationIdsRef = useRef<LaunchOptimizationId[]>([]);
-  const [gamescopeAutoSaveStatus, setGamescopeAutoSaveStatus] = useState<LaunchAutoSaveStatus>({
-    tone: 'idle',
-    label: 'Ready',
-  });
-  const gamescopeAutosaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const lastSavedGamescopeJsonRef = useRef<string>('null');
-  const [trainerGamescopeAutoSaveStatus, setTrainerGamescopeAutoSaveStatus] = useState<LaunchAutoSaveStatus>({
-    tone: 'idle',
-    label: 'Ready',
-  });
-  const trainerGamescopeAutosaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const lastSavedTrainerGamescopeJsonRef = useRef<string>('null');
-  const [mangoHudAutoSaveStatus, setMangoHudAutoSaveStatus] = useState<LaunchAutoSaveStatus>({
-    tone: 'idle',
-    label: 'Ready',
-  });
-  const mangoHudAutosaveTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const lastSavedMangoHudJsonRef = useRef<string>('null');
-  const hasExistingSavedProfileRef = useRef(false);
-  const pendingLaunchPresetRef = useRef<string | null>(null);
-  const profileRef = useRef(profile);
-  profileRef.current = profile;
-  const selectedProfileRef = useRef(selectedProfile);
-  selectedProfileRef.current = selectedProfile;
-
-  const onAfterRollbackRef = useRef(options.onAfterRollback);
-  onAfterRollbackRef.current = options.onAfterRollback;
-
   const { catalog, loading: catalogLoading } = useLaunchOptimizationCatalog();
   const optionsById = useMemo(() => (catalog ? buildOptionsById(catalog.entries) : {}), [catalog]);
   const conflictMatrix = useMemo(() => (catalog ? buildConflictMatrix(catalog.entries) : {}), [catalog]);
 
-  /** Serializes launch-optimization disk IPC so autosave / flush / preset saves cannot clobber each other. */
-  const launchProfileWriteChainRef = useRef<Promise<unknown>>(Promise.resolve());
-  const enqueueLaunchProfileWrite = useCallback(<T>(fn: () => Promise<T>): Promise<T> => {
-    const run: Promise<T> = launchProfileWriteChainRef.current.then(() => fn());
-    launchProfileWriteChainRef.current = run.then(
-      () => undefined,
-      () => undefined
-    );
-    return run;
-  }, []);
-
-  const syncProfileMetadata = useCallback(async (name: string, currentProfile: GameProfile) => {
-    const settings = await callCommand<AppSettingsData>('settings_load');
-    await callCommand('settings_save', {
-      data: toSettingsSaveRequest({ ...settings, last_used_profile: name }),
-    });
-
-    const recentFiles = await callCommand<RecentFilesData>('recent_files_load');
-    await callCommand('recent_files_save', {
-      data: {
-        game_paths: mergeRecentPaths(recentFiles.game_paths, currentProfile.game.executable_path),
-        trainer_paths: mergeRecentPaths(recentFiles.trainer_paths, currentProfile.trainer.path),
-        dll_paths: currentProfile.injection.dll_paths.reduce(
-          (paths, dllPath) => mergeRecentPaths(paths, dllPath),
-          recentFiles.dll_paths
-        ),
-      } satisfies RecentFilesData,
-    });
-  }, []);
-
-  const loadFavorites = useCallback(async () => {
-    try {
-      const names = await callCommand<string[]>('profile_list_favorites');
-      setFavoriteProfiles(names);
-    } catch {
-      setFavoriteProfiles([]);
-    }
-  }, []);
-
-  const toggleFavorite = useCallback(
-    async (name: string, favorite: boolean) => {
-      try {
-        await callCommand('profile_set_favorite', { name, favorite });
-        await loadFavorites();
-      } catch (err) {
-        console.error('Failed to update profile favorite state', err);
-        throw err;
-      }
-    },
-    [loadFavorites]
-  );
-
-  const loadProfile = useCallback(
-    async (
-      name: string,
-      loadOptions?: {
-        /**
-         * Optional collection context. When provided (and non-empty), Rust will
-         * merge the collection's launch defaults via `effective_profile_with` and
-         * the returned profile reflects the merged view.
-         *
-         * EDITOR SAFETY INVARIANT: `ProfilesPage` callers MUST NOT pass this — the
-         * editor must always see the raw storage profile, otherwise edits would
-         * persist the merged view back to the profile TOML. Only the LaunchPage
-         * profile-selector path passes a collectionId.
-         */
-        collectionId?: string;
-        loadErrorContext?: string;
-        throwOnFailure?: boolean;
-      }
-    ) => {
-      const trimmed = name.trim();
-      if (!trimmed) {
-        setSelectedProfile('');
-        setProfileName('');
-        setProfile(createEmptyProfile());
-        setDirty(false);
-        lastSavedLaunchOptimizationIdsRef.current = [];
-        return;
-      }
-
-      setLoading(true);
-      setError(null);
-
-      const formatLoadError = (err: unknown) => (err instanceof Error ? err.message : String(err));
-
-      // Normalize collectionId: trim and drop empty so we never send a degenerate
-      // `""` to Rust (which would treat it as a missing-collection error). Undefined
-      // means "no collection context — return raw storage profile".
-      const collectionId = loadOptions?.collectionId?.trim() || undefined;
-
-      try {
-        const loaded = await callCommand<SerializedGameProfile>('profile_load', {
-          name: trimmed,
-          collectionId,
-        });
-        const normalized = normalizeProfileForEdit(loaded, optionsById);
-        setSelectedProfile(trimmed);
-        setProfileName(trimmed);
-        setProfile(normalized);
-        setDirty(false);
-        lastSavedLaunchOptimizationIdsRef.current = normalized.launch.optimizations.enabled_option_ids;
-        lastSavedGamescopeJsonRef.current = JSON.stringify(normalized.launch.gamescope ?? null);
-        lastSavedTrainerGamescopeJsonRef.current = JSON.stringify(normalized.launch.trainer_gamescope ?? null);
-        lastSavedMangoHudJsonRef.current = JSON.stringify(normalized.launch.mangohud ?? null);
-
-        try {
-          await syncProfileMetadata(trimmed, normalized);
-        } catch (syncErr) {
-          console.error('Failed to sync profile metadata (last-used profile, recent files)', syncErr);
-          setError(
-            `Profile loaded, but preferences sync failed: ${syncErr instanceof Error ? syncErr.message : String(syncErr)}`
-          );
-        }
-      } catch (err) {
-        const msg = formatLoadError(err);
-        const fullMsg = loadOptions?.loadErrorContext ? `${loadOptions.loadErrorContext}: ${msg}` : msg;
-        setError(fullMsg);
-        if (loadOptions?.throwOnFailure) {
-          throw fullMsg;
-        }
-      } finally {
-        setLoading(false);
-      }
-    },
-    [optionsById, syncProfileMetadata]
-  );
-
-  const refreshProfiles = useCallback(async () => {
-    try {
-      const names = await callCommand<string[]>('profile_list');
-      setProfiles(names);
-
-      if (names.length === 0) {
-        setSelectedProfile('');
-        setProfileName('');
-        setProfile(createEmptyProfile());
-        setDirty(false);
-        lastSavedLaunchOptimizationIdsRef.current = [];
-        return;
-      }
-
-      if (selectedProfile && names.includes(selectedProfile)) {
-        return;
-      }
-
-      if (options.autoSelectFirstProfile ?? true) {
-        await loadProfile(names[0]);
-      }
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-    }
-  }, [loadProfile, options.autoSelectFirstProfile, selectedProfile]);
-
-  const selectProfile = loadProfile;
-
-  const finalizeProfileDeletion = useCallback(
-    async (name: string) => {
-      const settings = await callCommand<AppSettingsData>('settings_load');
-      if (settings.last_used_profile === name) {
-        await callCommand('settings_save', {
-          data: toSettingsSaveRequest({ ...settings, last_used_profile: '' }),
-        });
-      }
-      const names = await callCommand<string[]>('profile_list');
-      setProfiles(names);
-      void loadFavorites();
-
-      if (names.length === 0) {
-        setSelectedProfile('');
-        setProfileName('');
-        setProfile(createEmptyProfile());
-        setDirty(false);
-        lastSavedLaunchOptimizationIdsRef.current = [];
-        return;
-      }
-
-      await loadProfile(names[0], {
-        loadErrorContext: 'Profile deleted, but loading the next profile failed',
-      });
-    },
-    [loadFavorites, loadProfile]
-  );
-
-  const hydrateProfile = useCallback(
-    (name: string, nextProfile: GameProfile) => {
-      const trimmedName = name.trim();
-      if (!trimmedName) {
-        setError('Profile name is required.');
-        return;
-      }
-
-      const normalizedProfile = normalizeProfileForEdit(nextProfile, optionsById);
-
-      setSelectedProfile(profiles.includes(trimmedName) ? trimmedName : '');
-      setProfileName(trimmedName);
-      setProfile(normalizedProfile);
-      lastSavedLaunchOptimizationIdsRef.current = normalizedProfile.launch.optimizations.enabled_option_ids;
-      setDirty(true);
-      setError(null);
-    },
-    [optionsById, profiles]
-  );
-
-  const updateProfile = useCallback((updater: (current: GameProfile) => GameProfile) => {
-    setProfile((current) => updater(current));
-    setDirty(true);
-  }, []);
-
-  /** Like `updateProfile`, but only marks dirty when the profile has never been saved.
-   *  Use for fields that are auto-saved (gamescope, mangohud) so the Profiles page
-   *  does not show "Unsaved changes" after auto-save completes. */
-  const updateLaunchSetting = useCallback((updater: (current: GameProfile) => GameProfile) => {
-    setProfile((current) => updater(current));
-    setDirty((currentDirty) => currentDirty || !hasExistingSavedProfileRef.current);
-  }, []);
-
-  const hasExistingSavedProfile = useMemo(() => {
-    const trimmedName = profileName.trim();
-    return (
-      trimmedName.length > 0 &&
-      selectedProfile.trim().length > 0 &&
-      trimmedName === selectedProfile &&
-      profiles.includes(trimmedName)
-    );
-  }, [profileName, profiles, selectedProfile]);
-
-  hasExistingSavedProfileRef.current = hasExistingSavedProfile;
-
-  /** Clears any debounced launch-optimizations timer and persists pending edits immediately (same IPC as autosave). */
-  const flushPendingLaunchOptimizationsSave = useCallback(
-    async (nameForSave: string): Promise<void> => {
-      if (launchOptimizationsAutosaveTimerRef.current !== null) {
-        clearTimeout(launchOptimizationsAutosaveTimerRef.current);
-        launchOptimizationsAutosaveTimerRef.current = null;
-      }
-
-      if (!hasExistingSavedProfileRef.current) {
-        return;
-      }
-
-      const current = profileRef.current;
-      const method = resolveLaunchMethod(current);
-      if (method !== 'proton_run' && method !== 'steam_applaunch') {
-        return;
-      }
-
-      const trimmed = nameForSave.trim();
-      if (!trimmed) {
-        return;
-      }
-
-      const currentIds = current.launch.optimizations.enabled_option_ids;
-      if (areLaunchOptimizationIdsEqual(currentIds, lastSavedLaunchOptimizationIdsRef.current)) {
-        return;
-      }
-
-      await enqueueLaunchProfileWrite(async () => {
-        await callCommand('profile_save_launch_optimizations', {
-          name: trimmed,
-          optimizations: {
-            enabled_option_ids: [...currentIds],
-          },
-        });
-        lastSavedLaunchOptimizationIdsRef.current = [...currentIds];
-      });
-    },
-    [enqueueLaunchProfileWrite]
-  );
-
-  const toggleLaunchOptimization = useCallback(
-    (optionId: LaunchOptimizationId, nextEnabled: boolean) => {
-      const result = applyLaunchOptimizationToggle(
-        profileRef.current,
-        optionId,
-        nextEnabled,
-        optionsById,
-        conflictMatrix
-      );
-      if (!result.ok) {
-        setLaunchOptimizationsStatus({
-          tone: 'warning',
-          label: 'Conflicting option blocked',
-          detail: `Disable ${result.conflictLabels.join(' or ')} before enabling ${optionsById[optionId]?.label ?? optionId}.`,
-        });
-        return;
-      }
-
-      setProfile(result.profile);
-      setDirty((currentDirty) => currentDirty || !hasExistingSavedProfile);
-    },
-    [conflictMatrix, hasExistingSavedProfile, optionsById]
-  );
-
-  const switchLaunchOptimizationPreset = useCallback(
-    async (presetName: string): Promise<void> => {
-      const trimmedName = profileName.trim();
-      const key = presetName.trim();
-      const requestProfileName = trimmedName;
-
-      if (!trimmedName || !hasExistingSavedProfile || !key) {
-        return;
-      }
-
-      const presets = profileRef.current.launch.presets ?? {};
-      if (!presets[key]) {
-        setLaunchOptimizationsStatus({
-          tone: 'warning',
-          label: 'Unknown preset',
-          detail: `No launch optimization preset named "${key}" is defined in this profile.`,
-        });
-        return;
-      }
-
-      const method = resolveLaunchMethod(profileRef.current);
-      if (method !== 'proton_run' && method !== 'steam_applaunch') {
-        return;
-      }
-
-      try {
-        await flushPendingLaunchOptimizationsSave(trimmedName);
-      } catch (err) {
-        setDirty(true);
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to save',
-          detail: message,
-        });
-        return;
-      }
-
-      launchOptimizationsAutosaveTimerRef.current = null;
-
-      const currentAfterFlush = profileRef.current;
-      const presetsMap = currentAfterFlush.launch.presets ?? {};
-      const targetSection = presetsMap[key];
-      if (!targetSection) {
-        setLaunchOptimizationsStatus({
-          tone: 'warning',
-          label: 'Unknown preset',
-          detail: `No launch optimization preset named "${key}" is defined in this profile.`,
-        });
-        return;
-      }
-
-      // IDs for the preset we are switching *to* (not the previously active one). The backend
-      // currently ignores enabled_option_ids when switch_active_preset is set, but the payload
-      // should still match the target preset. We reuse this snapshot after IPC instead of
-      // profileRef.current, which may not reflect disk if state moves elsewhere.
-      const targetPresetIds = normalizeLaunchOptimizationIds(targetSection.enabled_option_ids, optionsById);
-
-      setError(null);
-      pendingLaunchPresetRef.current = key;
-
-      try {
-        await enqueueLaunchProfileWrite(async () => {
-          await callCommand('profile_save_launch_optimizations', {
-            name: trimmedName,
-            optimizations: {
-              enabled_option_ids: [...targetPresetIds],
-              switch_active_preset: key,
-            },
-          });
-        });
-
-        if (pendingLaunchPresetRef.current !== key || selectedProfileRef.current.trim() !== requestProfileName) {
-          return;
-        }
-
-        setProfile((current) => ({
-          ...current,
-          launch: {
-            ...current.launch,
-            active_preset: key,
-            optimizations: { enabled_option_ids: targetPresetIds },
-          },
-        }));
-        lastSavedLaunchOptimizationIdsRef.current = targetPresetIds;
-        setLaunchOptimizationsStatus({
-          tone: 'success',
-          label: 'Saved automatically',
-          detail: 'Active optimization preset updated.',
-        });
-      } catch (err) {
-        if (pendingLaunchPresetRef.current !== key || selectedProfileRef.current.trim() !== requestProfileName) {
-          return;
-        }
-
-        setDirty(true);
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to save',
-          detail: message,
-        });
-      } finally {
-        if (pendingLaunchPresetRef.current === key) {
-          pendingLaunchPresetRef.current = null;
-        }
-      }
-    },
-    [enqueueLaunchProfileWrite, flushPendingLaunchOptimizationsSave, hasExistingSavedProfile, optionsById, profileName]
-  );
-
-  const applyBundledOptimizationPreset = useCallback(
-    async (presetId: string): Promise<void> => {
-      const trimmedName = profileName.trim();
-      const pid = presetId.trim();
-      if (!trimmedName || !hasExistingSavedProfile || !pid) {
-        return;
-      }
-
-      const method = resolveLaunchMethod(profileRef.current);
-      if (method !== 'proton_run' && method !== 'steam_applaunch') {
-        return;
-      }
-
-      try {
-        await flushPendingLaunchOptimizationsSave(trimmedName);
-      } catch (err) {
-        setDirty(true);
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to save',
-          detail: message,
-        });
-        return;
-      }
-
-      launchOptimizationsAutosaveTimerRef.current = null;
-      setOptimizationPresetActionBusy(true);
-      setError(null);
-
-      try {
-        const updated = await enqueueLaunchProfileWrite(() =>
-          callCommand<SerializedGameProfile>('profile_apply_bundled_optimization_preset', {
-            name: trimmedName,
-            presetId: pid,
-          })
-        );
-        const normalized = normalizeProfileForEdit(updated, optionsById);
-        setProfile(normalized);
-        lastSavedLaunchOptimizationIdsRef.current = normalized.launch.optimizations.enabled_option_ids;
-        setLaunchOptimizationsStatus({
-          tone: 'success',
-          label: 'Bundled preset applied',
-          detail: 'Preset saved under launch.presets and activated for this profile.',
-        });
-      } catch (err) {
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to apply bundled preset',
-          detail: message,
-        });
-      } finally {
-        setOptimizationPresetActionBusy(false);
-      }
-    },
-    [enqueueLaunchProfileWrite, flushPendingLaunchOptimizationsSave, hasExistingSavedProfile, optionsById, profileName]
-  );
-
-  const saveManualOptimizationPreset = useCallback(
-    async (presetDisplayName: string): Promise<void> => {
-      const trimmedName = profileName.trim();
-      const key = presetDisplayName.trim();
-      if (!trimmedName || !hasExistingSavedProfile) {
-        return;
-      }
-
-      if (!key) {
-        setLaunchOptimizationsStatus({
-          tone: 'warning',
-          label: 'Preset name required',
-          detail: 'Enter a name for the new preset.',
-        });
-        throw new Error('Preset name must not be empty');
-      }
-
-      if (key.startsWith('bundled/')) {
-        setLaunchOptimizationsStatus({
-          tone: 'warning',
-          label: 'Reserved name',
-          detail: 'Names starting with bundled/ are reserved for app-shipped GPU presets.',
-        });
-        throw new Error('Preset name must not start with bundled/');
-      }
-
-      const method = resolveLaunchMethod(profileRef.current);
-      if (method !== 'proton_run' && method !== 'steam_applaunch') {
-        return;
-      }
-
-      try {
-        await flushPendingLaunchOptimizationsSave(trimmedName);
-      } catch (err) {
-        setDirty(true);
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to save',
-          detail: message,
-        });
-        throw err;
-      }
-
-      launchOptimizationsAutosaveTimerRef.current = null;
-      const ids = normalizeLaunchOptimizationIds(
-        profileRef.current.launch.optimizations.enabled_option_ids,
-        optionsById
-      );
-      setOptimizationPresetActionBusy(true);
-      setError(null);
-
-      try {
-        const updated = await enqueueLaunchProfileWrite(() =>
-          callCommand<SerializedGameProfile>('profile_save_manual_optimization_preset', {
-            name: trimmedName,
-            presetName: key,
-            enabledOptionIds: ids,
-          })
-        );
-        const normalized = normalizeProfileForEdit(updated, optionsById);
-        setProfile(normalized);
-        lastSavedLaunchOptimizationIdsRef.current = normalized.launch.optimizations.enabled_option_ids;
-        setLaunchOptimizationsStatus({
-          tone: 'success',
-          label: 'Preset saved',
-          detail: `Saved as "${key}" and set as the active preset.`,
-        });
-      } catch (err) {
-        const message = formatInvokeError(err);
-        setError(message);
-        setLaunchOptimizationsStatus({
-          tone: 'error',
-          label: 'Failed to save preset',
-          detail: message,
-        });
-        throw err;
-      } finally {
-        setOptimizationPresetActionBusy(false);
-      }
-    },
-    [enqueueLaunchProfileWrite, flushPendingLaunchOptimizationsSave, hasExistingSavedProfile, optionsById, profileName]
-  );
-
-  const persistProfileDraft = useCallback(
-    async (name: string, draftProfile: GameProfile): Promise<PersistProfileDraftResult> => {
-      const trimmedName = name.trim();
-      if (!trimmedName) {
-        const message = 'Profile name is required.';
-        setError(message);
-        return { ok: false, error: message };
-      }
-
-      const validationError = validateProfileForSave(draftProfile);
-      if (validationError !== null) {
-        setError(validationError);
-        return { ok: false, error: validationError };
-      }
-
-      setSaving(true);
-      setError(null);
-
-      try {
-        const normalizedProfile = normalizeProfileForSave(draftProfile, optionsById);
-        await callCommand('profile_save', { name: trimmedName, data: normalizedProfile });
-        lastSavedLaunchOptimizationIdsRef.current = normalizedProfile.launch.optimizations.enabled_option_ids;
-        await syncProfileMetadata(trimmedName, normalizedProfile);
-        await refreshProfiles();
-        await loadProfile(trimmedName);
-        return { ok: true };
-      } catch (err) {
-        const message = err instanceof Error ? err.message : String(err);
-        setError(message);
-        return { ok: false, error: message };
-      } finally {
-        setSaving(false);
-      }
-    },
-    [loadProfile, optionsById, refreshProfiles, syncProfileMetadata]
-  );
-
-  const saveProfile = useCallback(async () => {
-    await persistProfileDraft(profileName, profile);
-  }, [persistProfileDraft, profile, profileName]);
-
-  /**
-   * Duplicates the profile identified by `sourceName` via the backend, then
-   * refreshes the profile list and auto-selects the newly created copy.
-   *
-   * The backend generates a unique name (e.g. "MyGame (Copy)") so the caller
-   * does not need to supply one. The `duplicating` state flag is true for
-   * the duration of the async operation, allowing the UI to show a spinner.
-   *
-   * @param sourceName - Name of the existing profile to duplicate.
-   */
-  const duplicateProfile = useCallback(
-    async (sourceName: string): Promise<void> => {
-      if (!sourceName.trim()) return;
-      setDuplicating(true);
-      setError(null);
-      try {
-        const result = await callCommand<SerializedDuplicateProfileResult>('profile_duplicate', {
-          name: sourceName,
-        });
-        await refreshProfiles();
-        await loadProfile(result.name);
-      } catch (err) {
-        const message = err instanceof Error ? err.message : String(err);
-        setError(message);
-      } finally {
-        setDuplicating(false);
-      }
-    },
-    [loadProfile, refreshProfiles]
-  );
-
-  /**
-   * Renames the profile identified by `oldName` to `newName` via the backend,
-   * then refreshes the profile list and auto-selects the renamed profile.
-   *
-   * The autosave timer is cancelled before the rename to prevent a debounced
-   * save from writing to the old filename while the rename is in progress.
-   *
-   * @param oldName - Current name of the profile to rename.
-   * @param newName - Desired new name for the profile.
-   */
-  const renameProfile = useCallback(
-    async (oldName: string, newName: string): Promise<RenameProfileResult> => {
-      if (!oldName.trim() || !newName.trim() || oldName.trim() === newName.trim()) {
-        return { ok: false, hadLauncher: false };
-      }
-
-      if (launchOptimizationsAutosaveTimerRef.current !== null) {
-        clearTimeout(launchOptimizationsAutosaveTimerRef.current);
-        launchOptimizationsAutosaveTimerRef.current = null;
-      }
-      if (gamescopeAutosaveTimerRef.current !== null) {
-        clearTimeout(gamescopeAutosaveTimerRef.current);
-        gamescopeAutosaveTimerRef.current = null;
-      }
-      if (trainerGamescopeAutosaveTimerRef.current !== null) {
-        clearTimeout(trainerGamescopeAutosaveTimerRef.current);
-        trainerGamescopeAutosaveTimerRef.current = null;
-      }
-      if (mangoHudAutosaveTimerRef.current !== null) {
-        clearTimeout(mangoHudAutosaveTimerRef.current);
-        mangoHudAutosaveTimerRef.current = null;
-      }
-
-      setRenaming(true);
-      setError(null);
-      try {
-        const hadLauncher = await callCommand<boolean>('profile_rename', {
-          oldName: oldName.trim(),
-          newName: newName.trim(),
-        });
-        await refreshProfiles();
-        await loadProfile(newName.trim());
-        await loadFavorites();
-        return { ok: true, hadLauncher };
-      } catch (err) {
-        setError(err instanceof Error ? err.message : String(err));
-        return { ok: false, hadLauncher: false };
-      } finally {
-        setRenaming(false);
-      }
-    },
-    [loadFavorites, loadProfile, refreshProfiles]
-  );
-
-  const confirmDelete = useCallback(async (name: string) => {
-    const trimmed = name.trim();
-    if (!trimmed) {
-      setError('Select a profile to delete.');
-      return;
-    }
-
-    setError(null);
-
-    try {
-      const launcherInfo = await callCommand<LauncherInfo>('check_launcher_for_profile', {
-        name: trimmed,
-      });
-
-      if (launcherInfo.script_exists || launcherInfo.desktop_entry_exists) {
-        setPendingDelete({ name: trimmed, launcherInfo });
-        return;
-      }
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-      return;
-    }
-
-    setPendingDelete({ name: trimmed, launcherInfo: null });
-  }, []);
-
-  const executeDelete = useCallback(async () => {
-    if (!pendingDelete) {
-      return;
-    }
-
-    const { name } = pendingDelete;
-    setPendingDelete(null);
-    setDeleting(true);
-    setError(null);
-
-    try {
-      await callCommand('profile_delete', { name });
-      await finalizeProfileDeletion(name);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : String(err));
-    } finally {
-      setDeleting(false);
-    }
-  }, [finalizeProfileDeletion, pendingDelete]);
-
-  const cancelDelete = useCallback(() => {
-    setPendingDelete(null);
-  }, []);
-
-  const fetchConfigHistory = useCallback(async (name: string, limit?: number): Promise<ConfigRevisionSummary[]> => {
-    setHistoryLoading(true);
-    setHistoryError(null);
-    try {
-      return await callCommand<ConfigRevisionSummary[]>('profile_config_history', {
-        name,
-        ...(limit !== undefined ? { limit } : {}),
-      });
-    } catch (err) {
-      const message = formatInvokeError(err);
-      setHistoryError(message);
-      throw message;
-    } finally {
-      setHistoryLoading(false);
-    }
-  }, []);
-
-  const fetchConfigDiff = useCallback(
-    async (name: string, revisionId: number, rightRevisionId?: number): Promise<ConfigDiffResult> => {
-      setHistoryLoading(true);
-      setHistoryError(null);
-      try {
-        return await callCommand<ConfigDiffResult>('profile_config_diff', {
-          name,
-          revisionId,
-          ...(rightRevisionId !== undefined ? { rightRevisionId } : {}),
-        });
-      } catch (err) {
-        const message = formatInvokeError(err);
-        setHistoryError(message);
-        throw message;
-      } finally {
-        setHistoryLoading(false);
-      }
-    },
-    []
-  );
-
-  const rollbackConfig = useCallback(
-    async (name: string, revisionId: number): Promise<ConfigRollbackResult> => {
-      setHistoryLoading(true);
-      setHistoryError(null);
-      try {
-        const result = await callCommand<ConfigRollbackResult>('profile_config_rollback', {
-          name,
-          revisionId,
-        });
-        await loadProfile(name, {
-          loadErrorContext: 'Rollback applied, but reloading the profile failed',
-          throwOnFailure: true,
-        });
-        onAfterRollbackRef.current?.(name);
-        return result;
-      } catch (err) {
-        const message = formatInvokeError(err);
-        setHistoryError(message);
-        throw message;
-      } finally {
-        setHistoryLoading(false);
-      }
-    },
-    [loadProfile]
-  );
-
-  const markKnownGood = useCallback(async (name: string, revisionId: number): Promise<void> => {
-    setHistoryLoading(true);
-    setHistoryError(null);
-    try {
-      await callCommand('profile_mark_known_good', { name, revisionId });
-    } catch (err) {
-      const message = formatInvokeError(err);
-      setHistoryError(message);
-      throw message;
-    } finally {
-      setHistoryLoading(false);
-    }
-  }, []);
+  const setLastSavedProfileSnapshotRef = useRef<(profile: GameProfile) => void>(() => {});
+  const clearAutosaveTimersRef = useRef<() => void>(() => {});
+
+  const crud = useProfileCrud({
+    optionsById,
+    autoSelectFirstProfile: options.autoSelectFirstProfile ?? true,
+    setLastSavedProfileSnapshot: (nextProfile) => setLastSavedProfileSnapshotRef.current(nextProfile),
+    clearAutosaveTimers: () => clearAutosaveTimersRef.current(),
+  });
+
+  const launchAutosave = useProfileLaunchAutosave({
+    profile: crud.profile,
+    profileName: crud.profileName,
+    selectedProfile: crud.selectedProfile,
+    hasExistingSavedProfile: crud.hasExistingSavedProfile,
+    optionsById,
+    conflictMatrix,
+    setProfile: crud.setProfile,
+    setDirty: crud.setDirty,
+    setError: crud.setError,
+  });
+
+  setLastSavedProfileSnapshotRef.current = launchAutosave.setLastSavedProfileSnapshot;
+  clearAutosaveTimersRef.current = launchAutosave.clearAutosaveTimers;
+
+  const history = useProfileHistory({
+    loadProfile: crud.loadProfile,
+    onAfterRollback: options.onAfterRollback,
+  });
+
+  const { refreshProfiles, loadFavorites, setError } = crud;
 
   useEffect(() => {
     void refreshProfiles().catch((err: unknown) => {
       setError(err instanceof Error ? err.message : String(err));
     });
     void loadFavorites();
-  }, [loadFavorites, refreshProfiles]);
-
-  useEffect(() => {
-    let cancelled = false;
-    void (async () => {
-      try {
-        const rows = await callCommand<BundledOptimizationPreset[]>('profile_list_bundled_optimization_presets');
-        if (!cancelled) {
-          setBundledOptimizationPresets(rows);
-        }
-      } catch {
-        if (!cancelled) {
-          setBundledOptimizationPresets([]);
-        }
-      }
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, []);
+  }, [loadFavorites, refreshProfiles, setError]);
 
   useEffect(() => {
     let active = true;
@@ -1367,301 +185,53 @@ export function useProfile(options: UseProfileOptions = {}): UseProfileResult {
       active = false;
       void unlistenPromise.then((unlisten) => unlisten());
     };
-  }, [loadFavorites, refreshProfiles]);
-
-  useEffect(() => {
-    const method = resolveLaunchMethod(profile);
-    const currentIds = profile.launch.optimizations.enabled_option_ids;
-
-    if (launchOptimizationsAutosaveTimerRef.current !== null) {
-      clearTimeout(launchOptimizationsAutosaveTimerRef.current);
-      launchOptimizationsAutosaveTimerRef.current = null;
-    }
-
-    if ((method !== 'proton_run' && method !== 'steam_applaunch') || !hasExistingSavedProfile) {
-      setLaunchOptimizationsStatus(buildLaunchOptimizationsStatus(method, hasExistingSavedProfile));
-      return;
-    }
-
-    if (areLaunchOptimizationIdsEqual(currentIds, lastSavedLaunchOptimizationIdsRef.current)) {
-      setLaunchOptimizationsStatus(buildLaunchOptimizationsStatus(method, true));
-      return;
-    }
-
-    setLaunchOptimizationsStatus({
-      tone: 'saving',
-      label: 'Saving...',
-      detail: 'Persisting only launch.optimizations for the current saved profile.',
-    });
-
-    const trimmedName = profileName.trim();
-    const normalizedIds = [...currentIds];
-    let cancelled = false;
-
-    launchOptimizationsAutosaveTimerRef.current = setTimeout(() => {
-      void (async () => {
-        try {
-          await enqueueLaunchProfileWrite(async () => {
-            await callCommand('profile_save_launch_optimizations', {
-              name: trimmedName,
-              optimizations: {
-                enabled_option_ids: normalizedIds,
-              },
-            });
-          });
-
-          if (cancelled) {
-            return;
-          }
-
-          lastSavedLaunchOptimizationIdsRef.current = normalizedIds;
-          setLaunchOptimizationsStatus({
-            tone: 'success',
-            label: 'Saved automatically',
-            detail: 'Only the Launch Optimizations section was written to disk.',
-          });
-        } catch (err) {
-          if (cancelled) {
-            return;
-          }
-
-          setDirty(true);
-          setLaunchOptimizationsStatus({
-            tone: 'error',
-            label: 'Failed to save',
-            detail: formatInvokeError(err),
-          });
-        }
-      })();
-    }, launchOptimizationsAutosaveDelayMs);
-
-    return () => {
-      cancelled = true;
-      if (launchOptimizationsAutosaveTimerRef.current !== null) {
-        clearTimeout(launchOptimizationsAutosaveTimerRef.current);
-        launchOptimizationsAutosaveTimerRef.current = null;
-      }
-    };
-  }, [enqueueLaunchProfileWrite, hasExistingSavedProfile, profile, profileName]);
-
-  // --- Gamescope auto-save effect (mirrors the optimization auto-save pattern above) ---
-  useEffect(() => {
-    const currentJson = JSON.stringify(profile.launch.gamescope ?? null);
-
-    if (gamescopeAutosaveTimerRef.current !== null) {
-      clearTimeout(gamescopeAutosaveTimerRef.current);
-      gamescopeAutosaveTimerRef.current = null;
-    }
-
-    if (!hasExistingSavedProfile) {
-      if (currentJson !== lastSavedGamescopeJsonRef.current) {
-        setGamescopeAutoSaveStatus({ tone: 'warning', label: 'Save profile first' });
-      }
-      return;
-    }
-
-    if (currentJson === lastSavedGamescopeJsonRef.current) {
-      setGamescopeAutoSaveStatus({ tone: 'idle', label: 'Ready' });
-      return;
-    }
-
-    setGamescopeAutoSaveStatus({ tone: 'saving', label: 'Saving...' });
-
-    const trimmedName = profileName.trim();
-    let cancelled = false;
-
-    gamescopeAutosaveTimerRef.current = setTimeout(() => {
-      void (async () => {
-        try {
-          await enqueueLaunchProfileWrite(async () => {
-            await callCommand('profile_save_gamescope_config', {
-              name: trimmedName,
-              config: profile.launch.gamescope ?? DEFAULT_GAMESCOPE_CONFIG,
-            });
-          });
-
-          if (cancelled) return;
-
-          lastSavedGamescopeJsonRef.current = currentJson;
-          setGamescopeAutoSaveStatus({ tone: 'success', label: 'Saved automatically' });
-        } catch (err) {
-          if (cancelled) return;
-          setDirty(true);
-          setGamescopeAutoSaveStatus({ tone: 'error', label: 'Failed to save', detail: formatInvokeError(err) });
-        }
-      })();
-    }, launchOptimizationsAutosaveDelayMs);
-
-    return () => {
-      cancelled = true;
-      if (gamescopeAutosaveTimerRef.current !== null) {
-        clearTimeout(gamescopeAutosaveTimerRef.current);
-        gamescopeAutosaveTimerRef.current = null;
-      }
-    };
-  }, [enqueueLaunchProfileWrite, hasExistingSavedProfile, profile, profileName]);
-
-  // --- Trainer gamescope auto-save effect ---
-  useEffect(() => {
-    const currentJson = JSON.stringify(profile.launch.trainer_gamescope ?? null);
-
-    if (trainerGamescopeAutosaveTimerRef.current !== null) {
-      clearTimeout(trainerGamescopeAutosaveTimerRef.current);
-      trainerGamescopeAutosaveTimerRef.current = null;
-    }
-
-    if (!hasExistingSavedProfile) {
-      if (currentJson !== lastSavedTrainerGamescopeJsonRef.current) {
-        setTrainerGamescopeAutoSaveStatus({ tone: 'warning', label: 'Save profile first' });
-      }
-      return;
-    }
-
-    if (currentJson === lastSavedTrainerGamescopeJsonRef.current) {
-      setTrainerGamescopeAutoSaveStatus({ tone: 'idle', label: 'Ready' });
-      return;
-    }
-
-    setTrainerGamescopeAutoSaveStatus({ tone: 'saving', label: 'Saving...' });
-
-    const trimmedName = profileName.trim();
-    let cancelled = false;
-
-    trainerGamescopeAutosaveTimerRef.current = setTimeout(() => {
-      void (async () => {
-        try {
-          await enqueueLaunchProfileWrite(async () => {
-            await callCommand('profile_save_trainer_gamescope_config', {
-              name: trimmedName,
-              config: profile.launch.trainer_gamescope ?? DEFAULT_GAMESCOPE_CONFIG,
-            });
-          });
-
-          if (cancelled) return;
-
-          lastSavedTrainerGamescopeJsonRef.current = currentJson;
-          setTrainerGamescopeAutoSaveStatus({ tone: 'success', label: 'Saved automatically' });
-        } catch (err) {
-          if (cancelled) return;
-          setDirty(true);
-          setTrainerGamescopeAutoSaveStatus({ tone: 'error', label: 'Failed to save', detail: formatInvokeError(err) });
-        }
-      })();
-    }, launchOptimizationsAutosaveDelayMs);
-
-    return () => {
-      cancelled = true;
-      if (trainerGamescopeAutosaveTimerRef.current !== null) {
-        clearTimeout(trainerGamescopeAutosaveTimerRef.current);
-        trainerGamescopeAutosaveTimerRef.current = null;
-      }
-    };
-  }, [enqueueLaunchProfileWrite, hasExistingSavedProfile, profile, profileName]);
-
-  // --- MangoHud auto-save effect (mirrors the optimization auto-save pattern above) ---
-  useEffect(() => {
-    const currentJson = JSON.stringify(profile.launch.mangohud ?? null);
-
-    if (mangoHudAutosaveTimerRef.current !== null) {
-      clearTimeout(mangoHudAutosaveTimerRef.current);
-      mangoHudAutosaveTimerRef.current = null;
-    }
-
-    if (!hasExistingSavedProfile) {
-      if (currentJson !== lastSavedMangoHudJsonRef.current) {
-        setMangoHudAutoSaveStatus({ tone: 'warning', label: 'Save profile first' });
-      }
-      return;
-    }
-
-    if (currentJson === lastSavedMangoHudJsonRef.current) {
-      setMangoHudAutoSaveStatus({ tone: 'idle', label: 'Ready' });
-      return;
-    }
-
-    setMangoHudAutoSaveStatus({ tone: 'saving', label: 'Saving...' });
-
-    const trimmedName = profileName.trim();
-    let cancelled = false;
-
-    mangoHudAutosaveTimerRef.current = setTimeout(() => {
-      void (async () => {
-        try {
-          await enqueueLaunchProfileWrite(async () => {
-            await callCommand('profile_save_mangohud_config', {
-              name: trimmedName,
-              config: profile.launch.mangohud ?? DEFAULT_MANGOHUD_CONFIG,
-            });
-          });
-
-          if (cancelled) return;
-
-          lastSavedMangoHudJsonRef.current = currentJson;
-          setMangoHudAutoSaveStatus({ tone: 'success', label: 'Saved automatically' });
-        } catch (err) {
-          if (cancelled) return;
-          setDirty(true);
-          setMangoHudAutoSaveStatus({ tone: 'error', label: 'Failed to save', detail: formatInvokeError(err) });
-        }
-      })();
-    }, launchOptimizationsAutosaveDelayMs);
-
-    return () => {
-      cancelled = true;
-      if (mangoHudAutosaveTimerRef.current !== null) {
-        clearTimeout(mangoHudAutosaveTimerRef.current);
-        mangoHudAutosaveTimerRef.current = null;
-      }
-    };
-  }, [enqueueLaunchProfileWrite, hasExistingSavedProfile, profile, profileName]);
-
-  const profileExists = useMemo(() => profiles.includes(profileName.trim()), [profileName, profiles]);
+  }, [loadFavorites, refreshProfiles, setError]);
 
   return {
-    profiles,
-    favoriteProfiles,
-    selectedProfile,
-    profileName,
-    profile,
-    dirty,
-    loading,
-    saving,
-    deleting,
-    duplicating,
-    renaming,
-    error,
-    profileExists,
-    pendingDelete,
-    launchOptimizationsStatus,
-    gamescopeAutoSaveStatus,
-    trainerGamescopeAutoSaveStatus,
-    mangoHudAutoSaveStatus,
-    historyLoading,
-    historyError,
-    setProfileName,
-    selectProfile,
-    hydrateProfile,
-    updateProfile,
-    updateLaunchSetting,
-    toggleLaunchOptimization,
-    switchLaunchOptimizationPreset,
-    bundledOptimizationPresets,
-    applyBundledOptimizationPreset,
-    saveManualOptimizationPreset,
-    optimizationPresetActionBusy,
-    saveProfile,
-    duplicateProfile,
-    renameProfile,
-    persistProfileDraft,
-    confirmDelete,
-    executeDelete,
-    cancelDelete,
-    refreshProfiles,
-    toggleFavorite,
-    fetchConfigHistory,
-    fetchConfigDiff,
-    rollbackConfig,
-    markKnownGood,
+    profiles: crud.profiles,
+    favoriteProfiles: crud.favoriteProfiles,
+    selectedProfile: crud.selectedProfile,
+    profileName: crud.profileName,
+    profile: crud.profile,
+    dirty: crud.dirty,
+    loading: crud.loading,
+    saving: crud.saving,
+    deleting: crud.deleting,
+    duplicating: crud.duplicating,
+    renaming: crud.renaming,
+    error: crud.error,
+    profileExists: crud.profileExists,
+    pendingDelete: crud.pendingDelete,
+    launchOptimizationsStatus: launchAutosave.launchOptimizationsStatus,
+    gamescopeAutoSaveStatus: launchAutosave.gamescopeAutoSaveStatus,
+    trainerGamescopeAutoSaveStatus: launchAutosave.trainerGamescopeAutoSaveStatus,
+    mangoHudAutoSaveStatus: launchAutosave.mangoHudAutoSaveStatus,
+    historyLoading: history.historyLoading,
+    historyError: history.historyError,
+    setProfileName: crud.setProfileName,
+    selectProfile: crud.loadProfile,
+    hydrateProfile: crud.hydrateProfile,
+    updateProfile: crud.updateProfile,
+    updateLaunchSetting: crud.updateLaunchSetting,
+    toggleLaunchOptimization: launchAutosave.toggleLaunchOptimization,
+    switchLaunchOptimizationPreset: launchAutosave.switchLaunchOptimizationPreset,
+    bundledOptimizationPresets: launchAutosave.bundledOptimizationPresets,
+    applyBundledOptimizationPreset: launchAutosave.applyBundledOptimizationPreset,
+    saveManualOptimizationPreset: launchAutosave.saveManualOptimizationPreset,
+    optimizationPresetActionBusy: launchAutosave.optimizationPresetActionBusy,
+    saveProfile: crud.saveProfile,
+    duplicateProfile: crud.duplicateProfile,
+    renameProfile: crud.renameProfile,
+    persistProfileDraft: crud.persistProfileDraft,
+    confirmDelete: crud.confirmDelete,
+    executeDelete: crud.executeDelete,
+    cancelDelete: crud.cancelDelete,
+    refreshProfiles: crud.refreshProfiles,
+    toggleFavorite: crud.toggleFavorite,
+    fetchConfigHistory: history.fetchConfigHistory,
+    fetchConfigDiff: history.fetchConfigDiff,
+    rollbackConfig: history.rollbackConfig,
+    markKnownGood: history.markKnownGood,
     catalog,
     catalogLoading,
   };
